puzzle four
My initial code for this was a disaster. I have cleaned it up
significantly though the algorithms are the same. I was happy to get the
correct answer to both parts on the first try without using the test
data first. While this proves my mental model of my program was correct
my code looks massively overcomplicated compared to other solutions. I'm
guessing I missed an aspect to the problem or used inferior data
structures.