puzzle five
This one took longer than it should have. I had a couple of stray bytes
on the end of my file. :/ This is the first puzzle I had to resort to
running the test data to confirm the algorithm I used.

As to the algorithm this is the worst performing puzzle so far. Looking
at the reddit thread a nice way to do this is using a couple of
stacks/deques and moving chars from one to the other, removing matching
units on each side of the split. This only requires one pass.