puzzle nine
The test data worked first try though the actual data uncovered a bug in
the rotate 7 part of my code. Eventually just used a loop to do this.
I did part one using a vec and keeping a current index. This worked
though became unfeasible for part two. Switched to a deque and kept the
current index at the beginning. This is still not the best performance
wise though was good enough.