puzzle ten
Worked on the theory that when the text appeared the spread of the
points would be a minimum. I printed out min/max for x/y for each second
and just used less to page through till I found a minimum of 61 x 9.
Then plotted this out as ASCII and used less again as an image viewer to
find the text. Part two was easy. Spent some time cleaning up my code
and isolating the text. I initially printed the text as 0s and 1s and
this caused me to misread the last letter.