puzzle eleven
Part one was simple. Only trouble I had was getting the power equation
wrong three times before finally getting it. For part two I was able to
get this by brute force by trying squares of 1..300. My answer was of
size 9 and this was found fairly easily before progress stalled. It
wasn't an exhaustive search though so I switched to doing an expanding
square approach. Take a square, add the edges on both sides, add the
corner and repeat. This allows a search of the full solution space in
less than a second. So far this is the longest running solution of any.